Changed to use docker

+++ docker/httpd/build.sh
@@ -1,4 +1,11 @@
 #!/bin/bash
+if [ -z "$1" ] || [ -z "$2" ] || [ -z "$3" ]; then
+  errorMessage="Build requires a sitename and port as follows: ./build sitename 8080 volumedirectory (/var/docker_volumes/sitename)"
+  echo $errorMessage
+  echo "SYNTAX: ./build.sh <sitename> <port> <volumedirectory>"
+  exit 1;
+fi
+
 rm -rf $3
 docker kill $1
 docker rm $1
